+++ src/molecules/pagination/Pagination.js
@@ -1,38 +1,29 @@
 // @vue/component
 export default {
   props: {
-    page: {
+    /**
+     * Current page
+     */
+    currentPage: {
       type: Number,
       required: true
     },
-    totalSize: {
+    /**
+     * All pagers
+     */
+    numberOfPages: {
       type: Number,
       required: true
-    },
-    limit: {
-      type: Number,
-      default: 5
     }
   },
   data () {
     return {
       showFirst: false,
       showLast: false,
-      limitPerPage: this.limit || 5
+      visibleNumbers: 3
     }
   },
   computed: {
-    currentPage: {
-      get () {
-        return this.page
-      },
-      set (value) {
-        this.$emit('update:page', value)
-      }
-    },
-    numberOfPages () {
-      return Math.ceil(this.totalSize / this.limitPerPage)
-    },
     listOfPageNumbers () {
       return Array.from(Array(this.numberOfPages), (_, i) => i + 1)
     },
@@ -42,7 +33,7 @@ export default {
   },
   methods: {
     setLimitedPageNumber () {
-      if (this.numberOfPages <= this.limitPerPage) {
+      if (this.numberOfPages <= this.visibleNumbers) {
         this.showFirst = false
         this.showLast = false
 
@@ -50,24 +41,24 @@ export default {
       }
 
       if (
-        this.page <
-        this.limitPerPage - Math.floor(this.limitPerPage / 2) + 1
+        this.currentPage <
+        this.visibleNumbers - Math.floor(this.visibleNumbers / 2) + 1
       ) {
         this.showFirst = false
         this.showLast = true
 
-        return this.listOfPageNumbers.slice(0, this.limitPerPage)
+        return this.listOfPageNumbers.slice(0, this.visibleNumbers)
       }
 
       if (
-        this.numberOfPages - this.page <
-        this.limitPerPage - Math.floor(this.limitPerPage / 2) + 1
+        this.numberOfPages - this.currentPage <
+        this.visibleNumbers - Math.floor(this.visibleNumbers / 2) + 1
       ) {
         this.showFirst = true
         this.showLast = false
 
         return this.listOfPageNumbers.slice(
-          this.numberOfPages - this.limitPerPage
+          this.numberOfPages - this.visibleNumbers
         )
       }
 
@@ -75,12 +66,17 @@ export default {
       this.showLast = true
 
       return this.listOfPageNumbers.slice(
-        this.currentPage - Math.floor(this.limitPerPage / 2) - 1,
-        this.currentPage + Math.floor(this.limitPerPage / 2)
+        this.currentPage - Math.floor(this.visibleNumbers / 2) - 1,
+        this.currentPage + Math.floor(this.visibleNumbers / 2)
       )
     },
     setCurrentPage (value) {
       this.currentPage = value
+      /**
+       * Triggered when page is changed
+       * @type {Event}
+       */
+      this.$emit('update', value)
     }
   }
 }