+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Sidebar story based on Modal component
 - TransitionExpand for Collapse component
 - Disabled attribute for QuantityUpdate component
+- Pagination component
 
 ### Removed
 - -webkit- in transitions